Refactor pdfTeX version and timestamp in main.log

+++ download_retrieval_outputs.py
@@ -9,47 +9,71 @@ import shutil
 #                 # 'gl699': 'sphinx2',
 #                 }
 # targets = list(targets_dict.keys())
-targets = [
-    # 'gl411', 
-    'gl436', 'gl15A', 'gl876', 'gl832', 'gl752A', 
-        #    'gl338B', 
-        #    'gl905',
-           'gl849', 'gl725A', 'gl15B',
-           'gl687', 'gl880', 'gl725B', 
-        #    'gl699',
-           'gl1151',
-           ]
+
+targets_rv = {
+                'gl338B': 12.0,
+                'gl382' : 8.0,
+                'gl408' : 3.0,
+                'gl411' :-85.0,
+                'gl699' : -111.0,
+                'gl905' : -78.0,
+                'gl1286': 8.0,
+}
+targets = list(targets_rv.keys())
+
 targets_dict = dict(zip(targets, ['fc1']*len(targets)))
 
 
 base_path = pathlib.Path('/home/dario/phd/retrieval_base/')
-for target in targets:
-    # check if test_output dir exists
-    test_output = base_path / target / f'retrieval_outputs/{targets_dict[target]}/test_output'
-    if test_output.exists():
-        # remove test_output dir
-        # subprocess.run(f'rmdir {test_output}', shell=True, check=True)
-        shutil.rmtree(test_output)
-        
-    # download from snellius using scp -r
-    snellius_dir = f'/home/dgonzalezpi/retrieval_base/{target}/retrieval_outputs/{targets_dict[target]}/test_output'
-    local_dir = str(base_path / target / f'retrieval_outputs/{targets_dict[target]}/test_output')
+
+def download_run(target, run, cache=False):
     
+    # download from snellius using scp -r
+    snellius_dir = f'/home/dgonzalezpi/retrieval_base/{target}/retrieval_outputs/{run}/test_output'
+    local_dir = str(base_path / target / f'retrieval_outputs/{run}/test_output')
     print(f' Downloading {snellius_dir} to {local_dir}...')
-    try:
-        subprocess.run(f'scp -r dgonzalezpi@snellius.surf.nl:{snellius_dir} {local_dir}', shell=True, check=True)
-    # catch error and print message
-    except subprocess.CalledProcessError as e:
-        print(f' -> Error downloading {snellius_dir} to {local_dir}:\n{e}')
-        print(f' -> VPN must be disabled or set to NL!!')
+    
+    download_ok=False
+    if not cache:
+        try:
+            subprocess.run(f'scp -r dgonzalezpi@snellius.surf.nl:{snellius_dir} {local_dir}', shell=True, check=True)
+            print(f' Succesful download for {target} {run}!\n')
+            download_ok = True
+        # catch error and print message
+        except subprocess.CalledProcessError as e:
+            print(f' -> Error downloading {snellius_dir} to {local_dir}:\n{e}')
+            print(f' -> VPN must be disabled or set to NL!!')
+    else:
+        print(f' Using cached retrieval outputs for {target} {run}!\n')
+        download_ok = True
         
     try:
         # change working directory to target
-        subprocess.run(f'python retrieval_script.py -e -t {target}', shell=True, check=True, cwd=str(base_path / target))
+        subprocess.run(f'python retrieval_script.py -e --target {target} --run {run}', shell=True, check=True, cwd=str(base_path / target))
+        print(f' Succesful evaluation for {target} run {run}!\n')
+
     except subprocess.CalledProcessError as e:
         print(f' -> Error evaluating {target}:\n{e}')
+    
+    return download_ok
+
+cache = False
+try_runs = [f'fc{i}' for i in range(1,3)][::-1] # fc1, fc2
+
+for target in targets:
+    
         
-    print(f' Succesful download for {target}!\n')
+    for run in try_runs:
+        
+        # check if test_output dir exists
+        test_output = base_path / target / f'retrieval_outputs/{run}/test_output'
+        if test_output.exists() and not cache:
+            shutil.rmtree(test_output)
+            ok = download_run(target, run)
+        if ok:
+            break
+        
+    
     
     
 # run scripts to generate paper figures
